small fixes

+++ mainApp/src/main/java/logisticsmarshall/tqs/ua/model/Reputation.java
@@ -18,7 +18,7 @@ public class Reputation {
     private long id;
 
     @Column(name = "rating", nullable = false)
-    private Integer rating;
+    private long rating;
 
     @Column(name = "description", nullable = false)
     private String description;
